moved helpful helpers to loggers/archive.py with changes
